Rebalance affiliate members logos.
Previously, we unconditionally rendered all affiliate member logos
with a height of 50px. This was fine until recently, but now some
of the logos were *perceived* as bigger or smaller because of their
aspect ratio.

We rebalance the recent logos so that they are all *perceived* to
be the same size, even though some are now technically taller (but
narrower) than others.